CNN/requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.4.0
pandas>=1.3.0
scikit-learn>=1.0.0
h5py>=3.0.0
tqdm>=4.62.0
seaborn>=0.11.0
tensorboard>=2.10.0
PyYAML>=6.0
lmdb>=1.3.0